move md5 into encode package, and refactor other codes callling it
