+++ src/main/java/com/centaurstech/utils/encode/Md5.java
@@ -1,10 +1,10 @@
-package com.centaurstech.utils;
+package com.centaurstech.utils.encode;
 
 import java.math.BigInteger;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 
-@Deprecated
+
 public final class Md5 {
 
 	public static String digest(String plaintext) {
